+++ upcloud/firewall/testdata/upcloud_firewall_ruleset_s1.tf
@@ -1,0 +1,10 @@
+variable "ruleset_name" {
+  type = string
+}
+
+resource "upcloud_firewall_ruleset" "test" {
+  name                      = var.ruleset_name
+  description               = "Test firewall ruleset"
+  enabled                   = true
+  default_dns_rules_enabled = false
+}