+++ upcloud/firewall/testdata/upcloud_firewall_ruleset_s2.tf
@@ -1,0 +1,10 @@
+variable "ruleset_name" {
+  type = string
+}
+
+resource "upcloud_firewall_ruleset" "test" {
+  name                      = var.ruleset_name
+  description               = "Updated description"
+  enabled                   = false
+  default_dns_rules_enabled = true
+}