+++ upcloud/firewall/testdata/upcloud_firewall_ruleset_s3.tf
@@ -1,0 +1,8 @@
+variable "ruleset_name" {
+  type = string
+}
+
+resource "upcloud_firewall_ruleset" "test" {
+  name    = var.ruleset_name
+  enabled = false
+}